+++ packages/botonic-plugin-flow-builder/src/content-fields/index.ts
@@ -1,3 +1,4 @@
+import { FlowCustomConditionalV2 } from './custom-conditional/flow-custom-condition-v2'
 import { FlowAiAgent } from './flow-ai-agent'
 import { FlowAiAgentRouter } from './flow-ai-agent-router'
 import { FlowBotAction } from './flow-bot-action'
@@ -5,7 +6,6 @@ import { FlowCaptureUserInput } from './flow-capture-user-input'
 import { FlowCarousel } from './flow-carousel'
 import { FlowChannelConditional } from './flow-channel-conditional'
 import { FlowCountryConditional } from './flow-country-conditional'
-import { FlowCustomConditionalV2 } from './flow-custom-condition-v2'
 import { FlowCustomConditional } from './flow-custom-conditional'
 import { FlowGoToFlow } from './flow-go-to-flow'
 import { FlowHandoff } from './flow-handoff'
